Add showing shared data in home view
