+++ app/src/main/java/fi/metropolia/intellicircumstances/view/schedule/ScheduleResultsView.kt
@@ -23,6 +23,8 @@ import com.madrapps.plot.line.LineGraph
 import com.madrapps.plot.line.LinePlot
 import fi.metropolia.intellicircumstances.R
 import fi.metropolia.intellicircumstances.component.ShowAlertDialog
+import fi.metropolia.intellicircumstances.component.resultview.MeasurementTab
+import fi.metropolia.intellicircumstances.component.resultview.TabContent
 import fi.metropolia.intellicircumstances.database.Circumstance
 import fi.metropolia.intellicircumstances.ui.theme.Red100
 import fi.metropolia.intellicircumstances.ui.theme.Red200
@@ -148,151 +150,4 @@ fun ScheduleResultsView(
             }
         }
     )
-}
-
-@Composable
-private fun TabContent(
-    measurementTab: MeasurementTab,
-    circumstances: List<Circumstance>,
-    dateFormatter: (Long) -> String
-) {
-    var showInformation by rememberSaveable { mutableStateOf(false) }
-    var xCoordinate by rememberSaveable { mutableStateOf<Float?>(null) }
-    var dataPoint by rememberSaveable { mutableStateOf<DataPoint?>(null) }
-
-    val data = circumstances.mapNotNull {
-        when (measurementTab) {
-            MeasurementTab.HumidityTab -> it.humidity
-            MeasurementTab.PressureTab -> it.airPressure
-            MeasurementTab.TemperatureTab -> it.temperature
-        }
-    }
-    if (data.isNotEmpty()) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center,
-            modifier = Modifier
-                .fillMaxSize()
-        ) {
-            val unit = when (measurementTab) {
-                MeasurementTab.HumidityTab -> "%"
-                MeasurementTab.PressureTab -> "hPa"
-                MeasurementTab.TemperatureTab -> "\u00B0C"
-            }
-
-            Box {
-                ShowDataGraph(
-                    values = data,
-                    onSelectionStart = { showInformation = true },
-                    onSelectionEnd = { showInformation = false },
-                    onSelection = { x, dataP ->
-                        xCoordinate = x
-                        dataPoint = dataP
-                    }
-                )
-                if (showInformation && dataPoint != null) {
-                    ShowInformation(
-                        xCoordinate = xCoordinate!!,
-                        value = dataPoint!!.y,
-                        unit = unit,
-                        date = dateFormatter(circumstances[dataPoint!!.x.toInt()].time!!)
-                    )
-                }
-            }
-
-            Row(
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(top = 18.dp)
-            ) {
-                Text(
-                    text = String.format(stringResource(R.string.min_value), data.minOrNull(), unit),
-                    style = MaterialTheme.typography.subtitle2
-                )
-                Text(
-                    text = String.format(stringResource(R.string.max_value), data.maxOrNull(), unit),
-                    style = MaterialTheme.typography.subtitle2
-                )
-                Text(
-                    text = String.format(stringResource(R.string.avg_value), data.average(), unit),
-                    style = MaterialTheme.typography.subtitle2
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun ShowDataGraph(
-    values: List<Double>,
-    onSelectionStart: () -> Unit,
-    onSelectionEnd: () -> Unit,
-    onSelection: (Float, DataPoint) -> Unit,
-) {
-    LineGraph(
-        plot = LinePlot(
-            lines = listOf(
-                LinePlot.Line(
-                    dataPoints = values.mapIndexed { index, value ->
-                        DataPoint(index.toFloat(), value.toFloat())
-                    },
-                    connection = LinePlot.Connection(color = Red300),
-                    intersection = LinePlot.Intersection(color = Red500)
-                )
-            ),
-            yAxis = LinePlot.YAxis(steps = 10, roundToInt = false),
-            grid = LinePlot.Grid(color = Red100, steps = 8)
-        ),
-        onSelectionStart = onSelectionStart,
-        onSelectionEnd = onSelectionEnd,
-        onSelection = { x, dataPoints ->
-            onSelection(x, dataPoints.last())
-        },
-        modifier = Modifier
-            .fillMaxWidth()
-            .fillMaxHeight(0.75f)
-    )
-}
-
-@Composable
-private fun ShowInformation(xCoordinate: Float, value: Float, unit: String, date: String) {
-    var xSubtraction by rememberSaveable { mutableStateOf(0) }
-    Surface(
-        color = Red200,
-        elevation = 4.dp,
-        shape = RoundedCornerShape(8.dp),
-        modifier = Modifier
-            .height(60.dp)
-            .width(200.dp)
-            .onGloballyPositioned {
-                xSubtraction = it.size.width / 2
-            }
-            .graphicsLayer(translationX = xCoordinate - xSubtraction)
-    ) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center,
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(10.dp)
-        ) {
-            Text(
-                text = String.format(
-                    stringResource(R.string.value_at),
-                    value,
-                    unit,
-                    date
-                ),
-                style = MaterialTheme.typography.caption
-            )
-        }
-    }
-}
-
-private sealed class MeasurementTab(val tabName: Int) {
-    object HumidityTab: MeasurementTab(R.string.humid)
-    object PressureTab: MeasurementTab(R.string.pressure)
-    object TemperatureTab: MeasurementTab(R.string.temp)
 }
