+++ app/src/main/res/values/strings.xml
@@ -8,7 +8,7 @@
     <string name="cancel">Cancel</string>
     <string name="ok">OK</string>
     <string name="reset">Reset</string>
-    <string name="backto">Back to %s</string>
+    <string name="back_to">Back to %s</string>
     <string-array name="pref_keys">
         <item>dark_mode</item>
     </string-array>
@@ -31,6 +31,7 @@
     <string name="faq">FAQ</string>
     <string name="schedule_between">From %s to %s</string>
     <string name="community_data">Data shared by users</string>
+    <string name="date_error">Error formatting date</string>
 
     <!-- Spaces and properties -->
     <string name="spaces">Spaces</string>
@@ -95,6 +96,19 @@
     <string name="avg_value">Avg: %.2f %s</string>
     <string name="schedule_date">Showing schedule between %s and %s.</string>
     <string name="value_at">%.2f %s at %s</string>
+    <string name="schedule_already_shared_title">Results already shared</string>
+    <string name="schedule_already_shared_content">
+        Schedule results are already shared, so no action is required. The data is stored
+        anonymously and cannot be associated with you.
+    </string>
+    <string name="schedule_shared_title">Sharing succeeded</string>
+    <string name="schedule_shared_content">
+        Schedule results are now shared anonymously. You may view shared results at Home.
+    </string>
+    <string name="schedule_share_failed_title">Sharing failed</string>
+    <string name="schedule_share_failed_content">
+        Schedule results could not be shared due to an error. Please try again later.
+    </string>
 
     <!--settings-->
     <string name="settings">Settings</string>
@@ -111,17 +125,4 @@
     <string name="unnamed">Unnamed device</string>
     <string name="no_devices">No RuuviTag Devices.</string>
     <string name="contentdesc_delete_device">Using this button you can delete the RuuviTag device associated for %s\'s %s.</string>
-    <string name="schedule_already_shared_title">Results already shared</string>
-    <string name="schedule_already_shared_content">
-        Schedule results are already shared, so no action is required. The data is stored
-        anonymously and cannot be associated with you.
-    </string>
-    <string name="schedule_shared_title">Sharing succeeded</string>
-    <string name="schedule_shared_content">
-        Schedule results are now shared anonymously. You may view shared results at Home.
-    </string>
-    <string name="schedule_share_failed_title">Sharing failed</string>
-    <string name="schedule_share_failed_content">
-        Schedule results could not be shared due to an error. Please try again later.
-    </string>
 </resources>
